LICENSE: remove license application and add appendix
The Apache-2.0 appendix describes text that should go at the top of a
file to indicate that it is Apache-2.0 licensed, and to describe who the
license owner is.

Doing this seems to break github license detection, so remove it.

Additionally, add back the boilerplate appendix.

This should fix issue #1.

Signed-off-by: Sean Cross <sean@xobs.io>